feat: Add option for reps file in catalyst-toolbox
