+++ src/catalyst-toolbox/snapshot-lib/Cargo.toml
@@ -14,6 +14,7 @@ serde = { version = "1", features = ["derive"] }
 proptest = { workspace = true, branch = "master", optional = true }
 chain-addr = { path = "../../chain-libs/chain-addr", optional = true }
 test-strategy = { version = "0.2", optional = true }
+serde_json = "1.0"
 serde_test = { version = "1", optional = true }
 hex = { version = "0.4" }
 thiserror = "1.0"
@@ -28,7 +29,6 @@ rust_decimal_macros = "1"
 [dev-dependencies]
 serde_test = "1"
 test-strategy = "0.2"
-serde_json = "1.0"
 serde_yaml = "0.8.17"
 proptest = { workspace = true, branch = "master" }
 chain-addr = { path = "../../chain-libs/chain-addr" }
